Bug fix: Complete frames as soon as it is possible
It should not be necessary to wait for the subsequent sync word to
determine that the current frame is complete - this can be assured
simply back checking the expected frame size against the actual bytes.

This ensures that completed frames are made available immediately, and
avoids dropping a single frame from every binary stream that is
processed.